+++ src/controller/customerRecordsController.java
@@ -3,7 +3,7 @@ package controller;
 import helper.appointmentsQuery;
 import helper.customerRecordsQuery;
 import helper.loginPageQuery;
-import javafx.scene.control.Button;
+import javafx.scene.control.*;
 import model.customerModel;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -14,9 +14,6 @@ import javafx.fxml.Initializable;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Alert;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 
@@ -24,6 +21,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 /**
@@ -88,9 +86,26 @@ public class customerRecordsController implements Initializable {
      */
     public void deleteCustomer() throws SQLException {
         selectedCustomer = customerRecordsTable.getSelectionModel().getSelectedItem();
-        appointmentsQuery.deleteAllAppointments(selectedCustomer.getCustomerId());
-        customerRecordsQuery.delete(selectedCustomer.getCustomerId());
-        initTable();
+
+        if (selectedCustomer == null) {
+            displayCustomAlert("Customer not selected", "Please select a customer");
+        } else {
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Attention");
+            alert.setContentText("Are you sure you want to delete this customer?");
+            Optional<ButtonType> result = alert.showAndWait();
+
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                appointmentsQuery.deleteAllAppointments(selectedCustomer.getCustomerId());
+                customerRecordsQuery.delete(selectedCustomer.getCustomerId());
+                initTable();
+            }
+        }
+
+
+
+
+
     }
 
     /**
